+++ CHANGELOG.md
@@ -68,6 +68,7 @@ to docs, or any other relevant information.
   `typedActivity`.
 - **Experimental**: Signal definitions can now provide `TypeInfo` for converting Signal arguments on Client and
   Workflow callers and in Workflow handlers.
+- **Experimental**: Workflows can now use `TypeInfo` for Child Workflow inputs and results and continue-as-new inputs.
 
 ### Changed